+++ src/Program.cs
@@ -1,9 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Runtime.InteropServices;
-using System.Threading.Tasks;
-using CSSkinScrapper.UI;
 
 namespace CSSkinScrapper
 {
