+++ src/ScrapperImplemantations/Steam_Scrapper.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using CSSkinScrapper.Interop;
 
 namespace CSSkinScrapper.ScrapperImplemantations
 {
